requirements.txt
transformers
smart_open
neptune-client
pyarrow
lightning
neptune-pytorch-lightning